requirements.txt
pandas==2.3.3
openpyxl==3.1.5
exa-py==2.10.2
google-genai==1.69.0
python-dotenv==1.2.2
loguru==0.7.3
python-docx==1.2.0
langfuse==4.0.2
feedparser==6.0.12
fastapi==0.135.2
uvicorn==0.42.0
psycopg[binary]==3.3.3
